+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.12.2] - 8/24/2021
+
+### Added
+
+- Proper versions of `CollectionUtils.AddNonNull` that match `ICollection.Add` instead of `Enumerable.Concat`
+
+### Changed
+
+- Renamed old `CollectionUtils.AddNonNull` to `CollectionUtils.AppendNonNull`
+
 ## [2.12.1] - 8/24/2021
 
 ### Added